@@ -12,16 +12,18 @@
     "license": "MIT",
     "private": true,
     "dependencies": {
-        "@types/lodash": "^4.14.116",
+        "@types/lodash": "^4.14.155",
         "@types/minimist": "^1.2.0",
-        "@types/node": "8.*",
-        "@types/strip-json-comments": "^0.0.30",
-        "lodash": "^4.17.13",
-        "minimist": "^1.2.0",
-        "strip-json-comments": "^2.0.1",
-        "tar": "^4.4.6",
+        "@types/node": "^14.0.11",
+        "@types/request": "^2.48.5",
+        "@types/unzipper": "^0.10.3",
+        "fstream": "^1.0.12",
+        "lodash": "^4.17.15",
+        "minimist": "^1.2.5",
+        "strip-json-comments": "^3.1.0",
         "to-markdown": "^3.1.1",
-        "typescript": "^3.0.1",
+        "typescript": "^3.9.5",
+        "unzipper": "^0.10.11",
         "valid-url": "^1.0.9"
     },
     "defDependencies": {}
